+++ src/commands/functions-deletegcfartifacts.ts
@@ -20,7 +20,7 @@ function getConfirmationMessage(paths: string[]): string {
   return message;
 }
 
-export default new Command("functions:deletegcfartifacts")
+export const command = new Command("functions:deletegcfartifacts")
   .description(
     "Deletes all artifacts created by Google Cloud Functions on Google Container Registry."
   )