+++ src/commands/functions-list.ts
@@ -8,7 +8,7 @@ import * as backend from "../deploy/functions/backend";
 import { logger } from "../logger";
 import Table = require("cli-table");
 
-export default new Command("functions:list")
+export const command = new Command("functions:list")
   .description("list all deployed functions in your Firebase project")
   .before(requirePermissions, ["cloudfunctions.functions.list"])
   .action(async (options: Options) => {