+++ src/init/features/functions/index.ts
@@ -54,5 +54,5 @@ export async function doSetup(setup: any, config: any, options: Options) {
     default: "javascript",
     choices,
   });
-  return require("./" + language)(setup, config);
+  return require("./" + language).setup(setup, config);
 }